+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.1 (2015-01-28)
 
 - `merge` will not modify objects present on the target, rather it will create a new object or array and reassign the value. This fixes a bug whereby shared objects (for example, those on a parent class's prototype) were being mutated.
+- `requires` takes into account properties which are defined in the prototype chain
 
 ## 1.0.0 (2015-01-18)